Support array syntax
